Add CLI mode
In order to improve the help output and make it look
like the rest of helps in YADRO-KNS/obmc-yadro-cli,
add the `-c/--cli` option to switch the help format.

Add support for `help` argument in CLI mode. The argument
is equivalent to the `-h/--help` function.

Add `usage()` helper function and call it instead of
directly printing the usage, in CLI mode list all typically
available sensor types.

End-user-impact: Improved help output in CLI mode
Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>