Fix error message for invalid sensor type
Instead of raw DBus error, a user-friendly message
`No sensors of selected type are present` is printed
when a wrong sensor type is specified in the command line.
The message is followed by the standard `usage()` output
that in CLI mode lists all the typically available types.

End-user-impact: User-friendly error message on wrong sensor type
Signed-off-by: Alexander Amelkin <a.amelkin@yadro.com>